requirements.txt
flask
flask-sqlalchemy
flask-session
datetime
python-dotenv
Faker
wheel
gunicorn
